Database synchronisation bug
I have found that if I destroy the note using the prescribed method, the database becomes temporarily unsynchronised and the next keylist call triggered by the 'notedestroy' event still finds the supposedly destroyed note. This then causes an error to be thrown when a read is attempted with the key. The problem is fixed by calling SQNote.destroy({ where: { notekey: key } }); instead.